+++ ui/app/src/App.tsx
@@ -2,7 +2,6 @@ import React, { useEffect, useState } from 'react';
 import { DefaultPalette, IStackStyles, MessageBar, MessageBarType, Stack } from '@fluentui/react';
 import './App.scss';
 import { TopNav } from './components/shared/TopNav';
-import { Footer } from './components/shared/Footer';
 import { Routes, Route } from 'react-router-dom';
 import { RootLayout } from './components/root/RootLayout';
 import { WorkspaceProvider } from './components/workspaces/WorkspaceProvider';
@@ -12,21 +11,20 @@ import { Workspace } from './models/workspace';
 import { AppRolesContext } from './contexts/AppRolesContext';
 import { WorkspaceContext } from './contexts/WorkspaceContext';
 import { GenericErrorBoundary } from './components/shared/GenericErrorBoundary';
-import { NotificationsContext } from './contexts/NotificationsContext';
-import { Operation } from './models/operation';
-import { ResourceUpdate } from './models/resource';
+import { OperationsContext } from './contexts/OperationsContext';
+import { completedStates, Operation } from './models/operation';
 import { HttpMethod, ResultType, useAuthApiCall } from './hooks/useAuthApiCall';
 import { ApiEndpoint } from './models/apiEndpoints';
 import { CreateUpdateResource } from './components/shared/create-update-resource/CreateUpdateResource';
 import { CreateUpdateResourceContext } from './contexts/CreateUpdateResourceContext';
 import { CreateFormResource, ResourceType } from './models/resourceType';
+import { Footer } from './components/shared/Footer';
 
 export const App: React.FunctionComponent = () => {
   const [appRoles, setAppRoles] = useState([] as Array<string>);
   const [selectedWorkspace, setSelectedWorkspace] = useState({} as Workspace);
   const [workspaceRoles, setWorkspaceRoles] = useState([] as Array<string>);
   const [operations, setOperations] = useState([] as Array<Operation>);
-  const [resourceUpdates, setResourceUpdates] = useState([] as Array<ResourceUpdate>);
 
   const [createFormOpen, setCreateFormOpen] = useState(false);
   const [createFormResource, setCreateFormResource] = useState({ resourceType: ResourceType.Workspace } as CreateFormResource);
@@ -54,13 +52,13 @@ export const App: React.FunctionComponent = () => {
                 setCreateFormOpen(true);
               }
             }} >
-              <NotificationsContext.Provider value={{
+              <OperationsContext.Provider value={{
                 operations: operations,
                 addOperations: (ops: Array<Operation>) => {
                   let stateOps = [...operations];
                   ops.forEach((op: Operation) => {
                     let i = stateOps.findIndex((f: Operation) => f.id === op.id);
-                    if (i > 0) {
+                    if (i !== -1) {
                       stateOps.splice(i, 1, op);
                     } else {
                       stateOps.push(op);
@@ -68,18 +66,15 @@ export const App: React.FunctionComponent = () => {
                   });
                   setOperations(stateOps);
                 },
-                resourceUpdates: resourceUpdates,
-                addResourceUpdate: (r: ResourceUpdate) => {
-                  let updates = [...resourceUpdates];
-                  let i = updates.findIndex((f: ResourceUpdate) => f.resourceId === r.resourceId);
-                  if (i > 0) {
-                    updates.splice(i, 1, r);
-                  } else {
-                    updates.push(r);
-                  }
-                  setResourceUpdates(updates);
-                },
-                clearUpdatesForResource: (resourceId: string) => { let updates = [...resourceUpdates].filter((r: ResourceUpdate) => r.resourceId !== resourceId); setResourceUpdates(updates); }
+                dismissCompleted: () => {
+                  let stateOps = [...operations];
+                  stateOps.forEach((o:Operation) => {
+                    if(completedStates.includes(o.status)) {
+                      o.dismiss = true;
+                    }
+                  })
+                  setOperations(stateOps);
+                }
               }}>
                 <AppRolesContext.Provider value={{
                   roles: appRoles,
@@ -105,9 +100,9 @@ export const App: React.FunctionComponent = () => {
                           <Route path="/workspaces/:workspaceId//*" element={
                             <WorkspaceContext.Provider value={{
                               roles: workspaceRoles,
-                              setRoles: (roles: Array<string>) => { console.warn("Workspace roles", roles); setWorkspaceRoles(roles) },
+                              setRoles: (roles: Array<string>) => { console.info("Workspace roles", roles); setWorkspaceRoles(roles) },
                               workspace: selectedWorkspace,
-                              setWorkspace: (w: Workspace) => { console.warn("Workspace set", w); setSelectedWorkspace(w) },
+                              setWorkspace: (w: Workspace) => { console.info("Workspace set", w); setSelectedWorkspace(w) },
                               workspaceApplicationIdURI: selectedWorkspace.properties?.scope_id
                             }}>
                               <WorkspaceProvider />
@@ -121,7 +116,7 @@ export const App: React.FunctionComponent = () => {
                     </Stack.Item>
                   </Stack>
                 </AppRolesContext.Provider>
-              </NotificationsContext.Provider>
+              </OperationsContext.Provider>
             </CreateUpdateResourceContext.Provider>
           </MsalAuthenticationTemplate>
         } />